+++ ferbook/post/publish/index.php
@@ -93,5 +93,5 @@ $response["data"] = array(
     "postId" => $row->id
 );
 
-echo json_encode($response, JSON_UNESCAPED_UNICODE);
+echo json_encode($response, (float) JSON_UNESCAPED_UNICODE);
