+++ ferbook/search/index.php
@@ -170,4 +170,4 @@ if($query != "") {
     }
 }
 
-echo json_encode($response, JSON_UNESCAPED_UNICODE);
+echo json_encode($response, (float) JSON_UNESCAPED_UNICODE);
